requirements.txt
discord.py
python-dotenv
sqlalchemy
pynacl
yt-dlp
